backend/foodgram_backend/requirements.txt
asgiref==3.8.1
Django==3.2.16
djangorestframework==3.12.4
pillow==11.0.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
